delete item enhanced

+++ modules/delete_item.php
@@ -7,11 +7,14 @@
         header('Location: ../index.php');
     }
     $product_id=$_GET['id'];
-    $dlt_sql="UPDATE products SET availabilityStatus='deleted' WHERE productID='$product_id';";
-    mysqli_query($conn, $dlt_sql);
+    $dlt_sql="UPDATE products SET availabilityStatus='deleted' WHERE products.productID='$product_id' AND products.userID='$user_id';";
 
-    /* DELETE */
-    $dlt_sql="DELETE FROM cart_product WHERE productID='$product_id';";
-    mysqli_query($conn, $dlt_sql);
-    echo "<script>alert('Delete success!'); location.href='../supplier/index.php';</script>";
+    if (mysqli_query($conn, $dlt_sql)){
+        /* DELETE */
+        $dlt_sql="DELETE FROM cart_product WHERE productID='$product_id';";
+        mysqli_query($conn, $dlt_sql);
+        echo "<script>alert('Delete success!'); location.href='../index.php';</script>";
+    }else{
+        echo "<script>alert('Deletion failed!'); location.href='../index.php';</script>";
+    }
 ?>
